Filled in missing code and corrected info and names
